Sign out button now reaches onSignOut
What it actually does is pop up *another* button with the text 'Sign Out.'

It's in the wrong place and doesn't disappear when you log out.

That's bizarre, but it does return 204 No Content from the server,
which is the desired result. We can debug the appearance issues until
we get what we're looking for.

Responds to Issue #24.