User can successfully add a show listing
Modified show/events.js, api.js, and ui.js in order to
pass show data to the API.

Also modified /index.js and assets/scripts/index.js to point
to the /shows folder; before that happened, page refreshed on
submitting the form because nothing was reaching the event handler.

Console.log logs the show object and the object is the
correct format to pass to the API.

Request reaches API and returns 201 Created (success).

Confirmed that show data entered by user is in `shows` table in DB.

Closes Issue #14.