User can successfully update a show listing
Updated form in index.html to allow user to update a show listing

Updated code in /shows/ including events.js, api.js, and ui.js
to handle events and pass data to API.

An updated listing submitted via the form returns 200 OK from the API
as expected. Confirmed that data passed via the API is present in the
`shows` table in the database.

Closes Issue #37.