Show shows button now returns array of user's shows from API
Modified index.html to change form id to trigger event handler
to link 'show all shows' button with onShowAllShows function in
/shows/events.js.

Modified showAllShows in api.js to make GET request to API
to request a user's shows

Modified showShowsSuccess in ui.js to be sure we were
receiving the 'shows' array of objects

Added copious console.log statements.

Both api.js code and ui.js code get show object with array of user
shows. It's a little concerning that I cannot return the value of
the variable showShowList, but we're going to soldier on
to see if we can get things to display on the page via a handlebars
template now that we know we are returning the correct data
from the API.

Responds to Issue #28