Added moment.js library and code to format datetimes using library
Ran `npm install moment` to install moment.js, a library to format
datetimes in Javascript.

Details and additional documentation here:
http://momentjs.com/

Added code to require moment in /shows/ui.js
Added code to test moment.js on a timestamp from a particular show, using
`store.shows[0].show_time` to target a particular show's time property.

`console.log(moment(store.shows[0].show_time).format('MM/DD/YYYY h:mm a'))`

returns `01/01/2000 2:30 pm` in the console.

We're not going to want to use the date, but the next step will
be to have moment return just the time, and then figure out a way
to hand that over to show-listings.handlebars which populates
a user's show listings on the page.

Responds to Issue #68