Added starter /artworks files & began working on addArtworks modal
Added a folder, /scripts/artworks, and copied the code from
/scripts/shows into it as a base to start from.

Experimented with a number of modal approaches until
I came to one that worked.

No modals worked unless the .fullscreen_bg class was commented
out of index.scss. This is part of what came with the login modal,
but the fullscreen background would not tile, so it often cut off
features at the bottom, and wouldn't allow modals to pop up
so a user could edit them.  I'd like a fullscreen background back
but I may have to use another approach.

Responds to Issue #55. Plan to break up this issue into smaller,
feature based issues.